feat(#100): move web to internal
