+++ internal/web/app/app.go
@@ -16,11 +16,11 @@ import (
 	"github.com/Interhyp/metadata-service/internal/service/services"
 	"github.com/Interhyp/metadata-service/internal/service/trigger"
 	"github.com/Interhyp/metadata-service/internal/service/updater"
-	"github.com/Interhyp/metadata-service/web/controller/ownerctl"
-	"github.com/Interhyp/metadata-service/web/controller/repositoryctl"
-	"github.com/Interhyp/metadata-service/web/controller/servicectl"
-	"github.com/Interhyp/metadata-service/web/controller/webhookctl"
-	"github.com/Interhyp/metadata-service/web/server"
+	"github.com/Interhyp/metadata-service/internal/web/controller/ownerctl"
+	"github.com/Interhyp/metadata-service/internal/web/controller/repositoryctl"
+	"github.com/Interhyp/metadata-service/internal/web/controller/servicectl"
+	"github.com/Interhyp/metadata-service/internal/web/controller/webhookctl"
+	"github.com/Interhyp/metadata-service/internal/web/server"
 	auacorn "github.com/StephanHCB/go-autumn-acorn-registry"
 	"github.com/StephanHCB/go-backend-service-common/repository/logging"
 	"github.com/StephanHCB/go-backend-service-common/web/controller/healthctl"
