+++ internal/web/controller/servicectl/servicectl.go
@@ -14,8 +14,8 @@ import (
 	"github.com/Interhyp/metadata-service/acorns/errors/validationerror"
 	"github.com/Interhyp/metadata-service/acorns/service"
 	openapi "github.com/Interhyp/metadata-service/api/v1"
-	"github.com/Interhyp/metadata-service/web/middleware/jwt"
-	"github.com/Interhyp/metadata-service/web/util"
+	"github.com/Interhyp/metadata-service/internal/web/middleware/jwt"
+	"github.com/Interhyp/metadata-service/internal/web/util"
 	librepo "github.com/StephanHCB/go-backend-service-common/acorns/repository"
 	"github.com/StephanHCB/go-backend-service-common/web/util/media"
 	"github.com/go-chi/chi/v5"
