+++ internal/web/server/server.go
@@ -7,8 +7,8 @@ import (
 	"github.com/Interhyp/metadata-service/acorns/config"
 	"github.com/Interhyp/metadata-service/acorns/controller"
 	"github.com/Interhyp/metadata-service/acorns/repository"
-	"github.com/Interhyp/metadata-service/web/middleware"
-	"github.com/Interhyp/metadata-service/web/middleware/jwt"
+	"github.com/Interhyp/metadata-service/internal/web/middleware"
+	"github.com/Interhyp/metadata-service/internal/web/middleware/jwt"
 	auzerolog "github.com/StephanHCB/go-autumn-logging-zerolog"
 	"github.com/StephanHCB/go-autumn-logging-zerolog/loggermiddleware"
 	libcontroller "github.com/StephanHCB/go-backend-service-common/acorns/controller"
