+++ test/acceptance/util_setup_test.go
@@ -10,16 +10,16 @@ import (
 	"github.com/Interhyp/metadata-service/internal/service/mapper"
 	"github.com/Interhyp/metadata-service/internal/service/trigger"
 	"github.com/Interhyp/metadata-service/internal/service/updater"
+	"github.com/Interhyp/metadata-service/internal/web/app"
+	"github.com/Interhyp/metadata-service/internal/web/controller/ownerctl"
+	"github.com/Interhyp/metadata-service/internal/web/controller/repositoryctl"
+	"github.com/Interhyp/metadata-service/internal/web/controller/servicectl"
+	"github.com/Interhyp/metadata-service/internal/web/middleware/jwt"
+	"github.com/Interhyp/metadata-service/internal/web/server"
 	"github.com/Interhyp/metadata-service/test/acceptance/idpmock"
 	"github.com/Interhyp/metadata-service/test/acceptance/kafkamock"
 	"github.com/Interhyp/metadata-service/test/acceptance/metadatamock"
 	"github.com/Interhyp/metadata-service/test/acceptance/vaultmock"
-	"github.com/Interhyp/metadata-service/web/app"
-	"github.com/Interhyp/metadata-service/web/controller/ownerctl"
-	"github.com/Interhyp/metadata-service/web/controller/repositoryctl"
-	"github.com/Interhyp/metadata-service/web/controller/servicectl"
-	"github.com/Interhyp/metadata-service/web/middleware/jwt"
-	"github.com/Interhyp/metadata-service/web/server"
 	auacorn "github.com/StephanHCB/go-autumn-acorn-registry"
 	auconfigenv "github.com/StephanHCB/go-autumn-config-env"
 	librepo "github.com/StephanHCB/go-backend-service-common/acorns/repository"
