+++ test/contract/consumer/setup_ctr_test.go
@@ -2,7 +2,7 @@ package consumer
 
 import (
 	"github.com/Interhyp/metadata-service/internal/repository/config"
-	"github.com/Interhyp/metadata-service/web/app"
+	"github.com/Interhyp/metadata-service/internal/web/app"
 	auacorn "github.com/StephanHCB/go-autumn-acorn-registry"
 	auconfigenv "github.com/StephanHCB/go-autumn-config-env"
 	libconfig "github.com/StephanHCB/go-backend-service-common/repository/config"
